Enable overflow checking and fix overflow bugs in FuzzyTimeSpan and FuzzyUInt64

+++ src/Implementation/FuzzyUInt64.cs
@@ -3,6 +3,6 @@ namespace Fuzzy.Implementation
     sealed class FuzzyUInt64: Fuzzy<ulong>
     {
         public FuzzyUInt64(IFuzz fuzzy) : base(fuzzy) { }
-        public override ulong New() => fuzzy.UInt32() * fuzzy.UInt32();
+        public override ulong New() => (ulong)fuzzy.UInt32() * fuzzy.UInt32();
     }
 }
